pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "SPALF: Spectrally-Preconditioned Augmented Lagrangian on Stratified Feature Manifolds"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.4,<2.7",
    "transformers>=4.40,<4.49",
    "datasets>=3.0,<3.3",
    "numpy>=1.26,<2.2",
    "hydra-core>=1.3,<1.4",
    "omegaconf>=2.3,<2.4",
    "triton>=3.0,<3.2",
    "safetensors>=0.4,<0.5",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]
